+++ tests/cmd/test_hb.py
@@ -1,3 +1,19 @@
+################################################################
+#                                                              #
+#  This file is part of hermesbaby                             #
+#                       the software engineer's typewriter     #
+#                                                              #
+#  Copyright (c) 2024 Alexander Mann-Wahrenberg (basejumpa)    #
+#                                                              #
+#      https://github.com/basejumpa                            #
+#                                                              #
+#  License(s)                                                  #
+#                                                              #
+#  - MIT for contents used as software                         #
+#  - CC BY-SA-4.0 for contents used as method or otherwise     #
+#                                                              #
+################################################################
+
 """ CLI tests """
 
 import pytest