+++ CHANGELOG.rst
@@ -1,7 +1,42 @@
+..
+    ################################################################
+    #                                                              #
+    #  This file is part of hermesbaby                             #
+    #                       the software engineer's typewriter     #
+    #                                                              #
+    #  Copyright (c) 2024 Alexander Mann-Wahrenberg (basejumpa)    #
+    #                                                              #
+    #      https://github.com/basejumpa                            #
+    #                                                              #
+    #  License(s)                                                  #
+    #                                                              #
+    #  - MIT for contents used as software                         #
+    #  - CC BY-SA-4.0 for contents used as method or otherwise     #
+    #                                                              #
+    ################################################################
+
 Changelog
 #########
+
+Version 0.4
+***********
+
+- Add headers to all files
+
+Version 0.3
+***********
+
+- Add more packages
+
+
+Version 0.2
+***********
+
+- Add packages
+
 
 Version 0.1
 ***********
 
 - Omnium rerum principia parva sunt.
+